\name{annmap-package}
\alias{annmap-package}
\alias{annmap}
\docType{package}
\title{
  Provide access to the Annmap annotation database
}
\description{
  Annmap \url{http://annmap.picr.man.ac.uk} Is a genome annotation
  database and genome browser, based on the Google Maps API. The
  underlying annotation is derived from ENSEMBL
  (\url{http://www.ensembl.org}). Annmap also provides probe to genome
  mappings for Affymetrix Exon, Gene and Plus2 arrays.

  The annmap package makes the data in annmap available for use within
  R and BioConductor.
}
\details{
  \tabular{ll}{
    Package: \tab annmap\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2011-09-14\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Tim Yates\cr
  Maintainer: Tim Yates <tyates@picr.man.ac.uk>
}
\references{
  Yates T, Okoniewski MJ, Miller CJ. X:Map: annotation and visualization of genome structure for Affymetrix exon array analysis. Nucleic Acids Res. 2008 Jan;36(Database issue):D780-6. Epub 2007 Oct 11.\cr
  \url{http://nar.oxfordjournals.org/cgi/content/full/gkm779v1}
}
\keyword{ package }
\seealso{
 \code{\link{GenomicRanges}}
}
